+++ go.mod
@@ -7,13 +7,10 @@ toolchain go1.21.6
 require (
 	github.com/diamondburned/arikawa/v3 v3.3.5-0.20240122104110-dbc4ae8978dd
 	github.com/joho/godotenv v1.5.1
-	libdb.so/persist v0.0.0-20231219023831-5321494d3834
 )
 
 require (
-	github.com/fxamacker/cbor/v2 v2.5.0 // indirect
 	github.com/gorilla/schema v1.2.0 // indirect
 	github.com/gorilla/websocket v1.4.2 // indirect
-	github.com/x448/float16 v0.8.4 // indirect
 	golang.org/x/time v0.0.0-20210723032227-1f47c861a9ac // indirect
 )
